refactor and fix bug

+++ awscrawler/worker.py
@@ -113,7 +113,7 @@ def main():
     parser = argparse.ArgumentParser(description='Call Worker with arguments')
     parser.add_argument('--index', '-i', type=int, help='index of this machine in all this batch machines')
     parser.add_argument('--cookie', '-c', type=str, help='cookie for this machine')
-    parser.add_argument('--timeout', '-t', type=int, default=60, help='timeout for timehash to enqueue')
+    parser.add_argument('--timeout', '-t', type=float, default=60, help='timeout for timehash to enqueue')
     option = parser.parse_args()
     if option.index:
         obj = GetWorker(option.index)
